pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "agribottools"
version = "0.1.0"
description = "Utilities to convert datasets between Label Studio, YOLO and Ultralytics formats"
readme = "README.md"
requires-python = ">=3.8"
license = { file = "LICENSE.md" }
authors = [
  { name = "CNR Stiima" },
]
dependencies = [
    "opencv-python",
    "numpy",
    "python-dotenv",
    "tqdm",
    "PyYAML",
    "pillow",
    "requests",
    "gradio",
]

[tool.hatch.build.targets.wheel]
packages = ["src/cvtoolkit", "src/ui"]

[tool.hatch.build.targets.wheel.force-include]
"src/file_utils.py" = "file_utils.py"
